chore: separate encryption config from credstore config
